requirements.txt
flask
psycopg2-binary
requests
beautifulsoup4
python-dotenv
twilio
redis
gunicorn
Flask-Login
passlib
Flask-WTF
WTForms
celery
Flask-Mail
